+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Unreleased
 - Add function `StoreObject::resource_types` to get the types of resources that a store can hold
     - This can be used with a `StructureObject` as `StoreObject::try_from(structure_object).map(|store| store.resource_types())`
       which returns a `Result<&'static [ResourceType], StoreObjectConversionError>`
+- Add missing `StoreObject::Reactor` to the `seasonal-season-5` feature.
 
 0.20.1 (2024-01-09)
 ===================